【チャート参照つなぎこみ】20200522

+++ api/src/main/java/com/lifegraph/team20/models/Parent.java
@@ -2,19 +2,19 @@ package com.lifegraph.team20.models;
 
 public class Parent {
 
-	private long id;
+	private int id;
 
 
-	public Parent(long id) {
+	public Parent(int id) {
 		this.setId(id);
 
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
